extract get balance function from lib
